Adds ability to import associated occurrences
Associated occurrences can be imported in two ways:
* by mapping data to fields starting with associatedOccurrence:* (e.g., associatedOccurrence:identifer, associatedOccurrence:referenceUrl). This allows an easy way to use skeletal text file import to import associated occurrences, one per line, and potentially multiple per occurrence.
* by including data in JSON format in the associatedOccurrences field (now mappable)

During import for the former method, the data is packaged into JSON and stored in the associatedOccurrences field in the uploadspectemp table. If text already exists in that field, it is moved in to the verbatimText section of the JSON object. After specimens are transferred from uploadspectemp to omoccurrences, their associated occurrence data is added to the omoccurassociations table.

Before adding new associations, he code checks whether a particular associated occurrence already exists in the database based on the following:
* internalOccurrence: occid, occidAssociate, relationship (includes switching the occidAssociate and occid with the inverse relationship)
* externalOccurrence: occid, relationship, identifier (or resourceUrl if identifier is not set)
* genericObservation: occid, relationship, verbatimSciname

If the associated occurrence already exists, then it checks whether the rest of the data has changed, and if so updates the association.

If the occurrence is specified as an internalOccurrence with an occidAssociate, the code attempts to find that occurrence in the database or the current import to make sure that the correct occids are used. It looks first for an occid in omoccurrences that matches the identifier field. Failing that, it looks for a guid from guidoccurrences that matches the identifier field. Finally, it looks in uploadspectemp for an occurrenceID or dbpk that matches the identifier field, or a dbpk that matches the occidAssociate field. On finding any of these, it gets the occid of that record, and uses that for the occidAssociate field.

Finally, anything in the verbatimText part of the JSON object is put back into the associatedOccurrences field of the record in omoccurrences.